re: github issue https://github.com/Unidata/thredds/issues/1020
As noted in that issue, the realpath computed from a url suffix
must be done using TdsRequestedDataset.getLocationFromRequestPath().
However, when mocking, it must come from a different mechanism
(See Dap4Controller.getResourcePath()). This is because
I have not succeeded in geting Spring autowiring to work properly
when using controller mocking. This issue occurs in several DAP4
test programs, including TestServlet.